commands: move catalog command to separate pkg
