+++ command/cat/catalog_test.go
@@ -1,0 +1,13 @@
+package cat
+
+import (
+	"strings"
+	"testing"
+)
+
+func TestCatalogCommand_noTabs(t *testing.T) {
+	t.Parallel()
+	if strings.ContainsRune(New().Help(), '\t') {
+		t.Fatal("usage has tabs")
+	}
+}
