--- command/catalog_command_test.go
@@ -1,8 +1,0 @@
-package command
-
-import "testing"
-
-func TestCatalogCommand_noTabs(t *testing.T) {
-	t.Parallel()
-	assertNoTabs(t, new(CatalogCommand))
-}
